% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_aux.R
\name{args_to_string}
\alias{args_to_string}
\title{convert arguments and values of a function to a string to parse into other
functions}
\usage{
args_to_string(il)
}
\arguments{
\item{il}{list from \code{as.list(environment())} right after the function is
called.}
}
\value{
character
}
\description{
convert arguments and values of a function to a string to parse into other
functions
}
