% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{build_request_old}
\alias{build_request_old}
\title{build_request, OLD version}
\usage{
build_request_old(server, api_version, endpoint, ...)
}
\arguments{
\item{server}{character: Server. For WB internal use only}

\item{api_version}{character: API version}

\item{endpoint}{character: PIP API endpoint}

\item{...}{other parameters}
}
\value{
httr2 request
}
\description{
build_request, OLD version
}
