% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_available}
\alias{cache_available}
\alias{load_cache}
\alias{save_cache}
\title{Check if cache is available}
\usage{
cache_available(fhash)

load_cache(fhash)

save_cache(fhash, out, force = FALSE)
}
\arguments{
\item{fhash}{character: hash of calling function}

\item{out}{data to be cached}

\item{force}{logical. If TRUE force the creation of cache. Default is FALSE}
}
\value{
function hash
}
\description{
Check if cache is available
}
\section{Functions}{
\itemize{
\item \code{load_cache()}: Load cached data

\item \code{save_cache()}: Saves cache data

}}
\keyword{internal}
