% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{check_api}
\alias{check_api}
\title{Check internet connection and API status}
\usage{
check_api(api_version = "v1", server = NULL)
}
\arguments{
\item{api_version}{API version. Currently only \code{"v1"} is supported.}

\item{server}{One non-empty server name, or \code{NULL} for the production server.
For World Bank internal use only.}
}
\value{
character
}
\description{
Check internet connection and API status
}
\examples{
if (interactive()) {
check_api()
}
}
