% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{check_if_data}
\alias{check_if_data}
\title{check_if_data}
\usage{
check_if_data(df)
}
\arguments{
\item{df}{any. But will error if not of type data.frame}
}
\value{
Returns NULL. Will error if \code{df} is not class \code{data.frame}
}
\description{
check_if_data
}
