% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_is_data_frame}
\alias{check_is_data_frame}
\title{check_is_data_frame}
\usage{
check_is_data_frame(data)
}
\arguments{
\item{data}{any. But will error if it's not a data frame.#'}
}
\value{
Simply returns the data set if it passes, or errors if it fails.
}
\description{
A simple check, and error handling, for if something is written out as a 
non-data-frame erroneously.
}
