% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{check_multiple_outs}
\alias{check_multiple_outs}
\title{check_multiple_outs}
\usage{
check_multiple_outs(data_sets, paths)
}
\arguments{
\item{data_sets}{list. A list of data.frame objects.}

\item{paths}{list. A list of paths.}
}
\value{
Returns NULL. Will error if \code{length(data_sets) != length(paths)}
}
\description{
check_multiple_outs
}
