% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_full_paths}
\alias{create_full_paths}
\title{create_full_paths}
\usage{
create_full_paths(paths, root)
}
\arguments{
\item{paths}{character. A character string representing a directory location.}

\item{root}{character. What separates the directory from the file name. Defaults
to '/'}
}
\description{
A helper to create consistent full file paths when given a base directory and
some file paths.
}
