% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_cache}
\alias{delete_cache}
\title{Deletes content of the cache folder}
\usage{
delete_cache()
}
\value{
Side effect. Deletes files.
}
\description{
Deletes content of the cache folder
}
\examples{
if (interactive()) delete_cache()
}
