% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_aux.R
\name{display_aux}
\alias{display_aux}
\title{Display available auxiliary tables}
\usage{
display_aux(
  version = NULL,
  ppp_version = NULL,
  release_version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  simplify = TRUE,
  server = NULL,
  assign_tb = TRUE
)
}
\arguments{
\item{version}{Data version in \code{YYYYMMDD_PPP_XX_YY_IDENTITY} format, or
\code{NULL}. See \code{get_versions()}.}

\item{ppp_version}{One integer-valued numeric PPP year, a four-digit year
string, or \code{NULL}.}

\item{release_version}{A valid publication date in \code{YYYYMMDD} format, or
\code{NULL}.}

\item{api_version}{API version. Currently only \code{"v1"} is supported.}

\item{format}{Response format: \code{"rds"}, \code{"json"}, or \code{"csv"}.}

\item{simplify}{One non-missing logical value. If \code{TRUE} (the default),
return simplified data.}

\item{server}{One non-empty server name, or \code{NULL} for the production server.
For World Bank internal use only.}

\item{assign_tb}{logical: Whether to assign table to .pip env. Default is
TRUE}
}
\value{
invisible tibble with names of auxiliary tables
}
\description{
Display available auxiliary tables
}
\examples{
if (interactive()) {
display_aux()
}
}
