% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_gdp}
\alias{get_gdp}
\title{get_gdp()}
\usage{
get_gdp(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\description{
Returns a table of Growth Domestic Product (GDP) values used for
poverty and inequality
statistics
}
\examples{
\dontrun{
# Short hand to get gdp
get_gdp()
}
}
