% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_incgrp_coverage}
\alias{get_incgrp_coverage}
\title{get_incgrp_coverage}
\usage{
get_incgrp_coverage(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\arguments{
\item{version}{character: Data version. See \code{get_versions()}}

\item{api_version}{character: API version}

\item{format}{character: Response format either of c("rds", "json", "csv")}

\item{server}{character: Server. For WB internal use only}
}
\description{
Returns a table of survey coverage for low and lower-middle
income countries. If this coverage is less than 50\%, World level aggregate
statistics will not be computed.
}
\examples{
\dontrun{
# Short hand to get incgrp_coverage
get_incgrp_coverage()
}
}
