% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_pip_info}
\alias{get_pip_info}
\title{Get PIP info}
\usage{
get_pip_info(api_version = "v1", server = NULL)
}
\arguments{
\item{api_version}{API version. Currently only \code{"v1"} is supported.}

\item{server}{One non-empty server name, or \code{NULL} for the production server.
For World Bank internal use only.}
}
\value{
list
}
\description{
Get information about the API.
}
\examples{
if (interactive()) {
get_pip_info()
}
}
