% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_pop}
\alias{get_pop}
\title{get_pop}
\usage{
get_pop(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\description{
Returns a table of population values used for poverty and
inequality computations.
}
\examples{
\dontrun{
# Short hand to get pop
get_pop()
}
}
