% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_pop_region}
\alias{get_pop_region}
\title{get_pop_region}
\usage{
get_pop_region(
  version = NULL,
  ppp_version = NULL,
  release_version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\arguments{
\item{version}{character: Data version. See \code{get_versions()}}

\item{ppp_version}{ppp year to be used}

\item{release_version}{date when the data was published in YYYYMMDD format}

\item{api_version}{character: API version}

\item{format}{character: Response format either of c("rds", "json", "csv")}

\item{server}{character: Server. For WB internal use only}
}
\description{
Returns a table of total population by region-year. These values
are used for the computation of regional aggregate poverty statistics.
}
\examples{
\dontrun{
# Short hand to get pop_region
get_pop_region()
}
}
