% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_ppp}
\alias{get_ppp}
\title{get_ppp}
\usage{
get_ppp(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\description{
Returns a table of Purchasing Power Parity (PPP) values
used for poverty and inequality computations.
}
\examples{
\dontrun{
# Short hand to get ppp
get_ppp()
}
}
