% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_regions}
\alias{get_regions}
\title{get_regions}
\usage{
get_regions(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\arguments{
\item{version}{character: Data version. See \code{get_versions()}}

\item{api_version}{character: API version}

\item{format}{character: Response format}

\item{server}{character: Server. For WB internal use only}
}
\description{
Returns a table regional grouping used for computing aggregate
poverty statistics.
}
\examples{
\dontrun{
# Short hand to get regions
get_regions()
}
}
