% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_versions}
\alias{get_versions}
\title{Get versions}
\usage{
get_versions(api_version = "v1", server = NULL, simplify = TRUE)
}
\arguments{
\item{api_version}{API version. Currently only \code{"v1"} is supported.}

\item{server}{One non-empty server name, or \code{NULL} for the production server.
For World Bank internal use only.}

\item{simplify}{One non-missing logical value. If \code{TRUE} (the default),
return simplified data.}
}
\value{
tibble or list
}
\description{
Get available data versions.
}
\examples{
if (interactive()) {
get_versions()
}
}
