% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{health_check}
\alias{health_check}
\title{Check API status}
\usage{
health_check(api_version = "v1", server = NULL)
}
\arguments{
\item{api_version}{character: API version}

\item{server}{character: Server. For WB internal use only}
}
\description{
Check API status
}
\examples{
health_check()
}
