% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hijack}
\alias{hijack}
\title{hijack}
\usage{
hijack(fun, ...)
}
\arguments{
\item{fun}{function. A function you wish to stub out default options on.}

\item{...}{character. A named vector where the names correspond to function
arguments and the value corresponds to the default value you wish to insert.}
}
\value{
Returns a copy of the input, but with the new default values specified
by the \code{...}
}
\description{
Inspired from: https://www.r-bloggers.com/hijacking-r-functions-changing-default-arguments/
Helps reset the default arguments for functions
}
