% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_cached}
\alias{is_cached}
\alias{load_cache}
\alias{save_cache}
\title{Check if cache is available}
\usage{
is_cached(df = NULL, fhash = NULL)

load_cache(fhash)

save_cache(fhash, out, force = FALSE)
}
\arguments{
\item{df}{dataframe}

\item{fhash}{character: hash of calling function}

\item{out}{data to be cached}

\item{force}{logical. If TRUE force the creation of cache. Default is FALSE}
}
\value{
logical. whether the hash exist of the dataframe is cached.
}
\description{
Checks whether hash or dataframe is cached. Only of the two is
}
\section{Functions}{
\itemize{
\item \code{load_cache()}: Load cached data

\item \code{save_cache()}: Saves cache data

}}
\keyword{internal}
