% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_merge}
\alias{list_merge}
\title{list_merge}
\usage{
list_merge(list1, list2)
}
\arguments{
\item{list1}{named list.}

\item{list2}{named list.}
}
\value{
A list with all values from \code{list1} and any values from 
\code{list2} not specified by \code{list1}
}
\description{
A tool to combine two lists.
}
