% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{make_env}
\alias{make_env}
\title{make_env}
\usage{
make_env()
}
\value{
Creates the \code{.PipR_Env} environment and child environments
}
\description{
Internal function to set up the environments properly
}
