% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{map_0f1}
\alias{map_0f1}
\title{map_0f1}
\usage{
map_0f1(transform, outpath, writer)
}
\arguments{
\item{transform}{function. The transformation function to be applied.}

\item{outpath}{character. A character string representing a single path. The
transformed data set is written out to this path.}

\item{writer.}{The function used to write the data set to the target location.}
}
\value{
Returns the outpath.
}
\description{
map_0f1
}
