% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{map_0fM}
\alias{map_0fM}
\title{map_0fM}
\usage{
map_0fM(transform, outpaths, writer)
}
\arguments{
\item{transform}{function. The transformation function to be applied.}

\item{outpaths}{character. A character vector representing multiple output
paths, which each output set will be written to.}

\item{writer.}{The function used to write the data set to the target location.}
}
\value{
Returns the outpaths
}
\description{
map_0fM
}
