% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{map_1f1}
\alias{map_1f1}
\title{map_1f1}
\usage{
map_1f1(inpath, outpath, transform, reader, writer)
}
\arguments{
\item{inpath}{character. A character string representing a single path. The
raw data set is read in from this path.}

\item{outpath}{character. A character string representing a single path. The
transformed data set is written out to this path.}

\item{transform}{function. The transformation function to be applied.}

\item{writer.}{The function used to write the data set to the target location.}
}
\value{
Returns the outpath
}
\description{
map_1f1
}
