% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{map_1fM}
\alias{map_1fM}
\title{map_1fM}
\usage{
map_1fM(inpath, outpaths, transform, reader, writer)
}
\arguments{
\item{inpath}{character. A character string representing a single path. The
raw data set is read in from this path.}

\item{outpaths}{character. A character vector representing multiple output
paths, which each output set will be written to.}

\item{transform}{function. The transformation function to be applied.}

\item{writer.}{The function used to write the data set to the target location.}
}
\value{
Returns the outpaths
}
\description{
map_1fM
}
