% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{map_Mf1}
\alias{map_Mf1}
\title{map_Mf1}
\usage{
map_Mf1(inpaths, outpath, transform, reader, writer)
}
\arguments{
\item{inpaths}{character. A character string representing multiple input
paths. The raw data is read in from each path and stored as a list.}

\item{outpath}{character. A character string representing a single path. The
transformed data set is written out to this path.}

\item{transform}{function. The transformation function to be applied.}

\item{writer.}{The function used to write the data set to the target location.}
}
\value{
Returns the outpath
}
\description{
map_Mf1
}
