% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{map_MfM}
\alias{map_MfM}
\title{map_MfM}
\usage{
map_MfM(inpaths, outpaths, transform, reader, writer)
}
\arguments{
\item{inpaths}{character. A character string representing multiple input
paths. The raw data is read in from each path and stored as a list.}

\item{outpaths}{character. A character vector representing multiple output
paths, which each output set will be written to.}

\item{transform}{function. The transformation function to be applied.}

\item{writer.}{The function used to write the data set to the target location.}
}
\value{
Returns the outpaths
}
\description{
map_MfM
}
\details{
MfM is simply a linear mapping between the input and output. MfM 
is simply using `lapply` the transform against the input dataframes.
}
