% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nice_dir}
\alias{nice_dir}
\title{nice_dir}
\usage{
nice_dir(x, sep = "/")
}
\arguments{
\item{x}{character. A character string representing a directory location.}

\item{sep}{character. What separates the directory from the file name. Defaults
to '/'}
}
\description{
A little function to help deal with that fucking '/'
}
