% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_pipe}
\alias{parse_pipe}
\title{parse_pipe}
\usage{
parse_pipe(path)
}
\arguments{
\item{path}{}
}
\value{
Output is divided into two parts. Part one creates the environment 
\code{.PipR_Env} filled with the entries specified by config. Part two
returns a list object with two sections - \code{.$steps} and \code{.$checks}.
\code{parse_pipe()} also assigns the correct names to sections so \code{pipe()}
can read it properly.
}
\description{
Input path to PipR pipeline, creates a config environment that PipR steps use
by default, and outputs the list of PipR steps with neat names.
}
