% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipe}
\alias{pipe}
\title{pipe}
\usage{
pipe(path, only = NA, run_pipe = TRUE, run_checks = TRUE)
}
\arguments{
\item{path}{character. The path to your PipR pipeline}

\item{only}{integer. A vector representing which steps of the pipeline to run.}

\item{run_pipe}{logical. Controls whether to run the pipeline. Set to FALSE
to run checks only.}

\item{run_checks}{logical. Controls whether to run checks. Set to FALSE to
just run the pipeline without checking the output (not recommended).}
}
\value{
When \code{run_checks == TRUE}, \code{pipe} returns an s3 path 
where the check log is stored. When \code{run_checks == FALSE}, \code{pipe}
simply returns the list of files output by the model. Don't run with both
as \code{FALSE} what are you even doing?
}
\description{
The master function for PipR. Reads a yaml file and interprets it as 
instructions for transforming a series of data sets, creating snapshots of
those changes, and then applying checks to those snapshots for QA.
}
