% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_pipe}
\alias{read_pipe}
\title{read_pipe}
\usage{
read_pipe(path)
}
\arguments{
\item{path}{character. path to the pipeline file.}
}
\value{
Output is a list structured in the same way as the pipeline file.
}
\description{
Internal function just to read in the raw yaml.
}
