% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{remove_env}
\alias{remove_env}
\title{remove_env}
\usage{
remove_env()
}
\value{
Simply removes \code{.PipR_Env} from the search path and deletes
it from memory.
}
\description{
Clears any environments created when using PipR
}
