% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_cols}
\alias{rename_cols}
\title{rename columns in dataframe}
\usage{
rename_cols(df, oldnames, newnames)
}
\arguments{
\item{df}{data frame}

\item{oldnames}{character: old names}

\item{newnames}{character: new names}
}
\value{
data frame with new names
}
\description{
rename columns in dataframe
}
\keyword{internal}
