% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retry_after}
\alias{retry_after}
\title{retry_after}
\usage{
retry_after(resp)
}
\arguments{
\item{resp}{A httr response}
}
\value{
numeric
}
\description{
Helper function to determine how much time to wait before a new
query can be sent
}
