% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{run_all_checks}
\alias{run_all_checks}
\title{run_all_checks}
\usage{
run_all_checks(outpaths, checks, lcl)
}
\arguments{
\item{outpaths}{character. A character vector of one or more output paths 
produced by the pipeline. The data sets are read in from those paths for 
checks.}

\item{checks}{list. A character list, representing checks and possible input
parameters for that step.}

\item{lcl}{list. A list of the local configs at that step. Local configs are
specified in \code{run_step()} and replaced with global values as necessary.}
}
\description{
Runs all specified checks at a particular step.
}
