% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{run_check}
\alias{run_check}
\title{run_check}
\usage{
run_check(data, check, lcl, outpath)
}
\arguments{
\item{data}{data.frame. A data set to execute the check against.}

\item{check}{character. A character vector with the name of the function 
called, along with any possible input parameters.}

\item{lcl}{list. A list of local configs for that step.}

\item{outpath}{character. The path where the processed data was written to.}
}
\description{
run_check
}
