% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{run_cli}
\alias{run_cli}
\title{whether or not to run cli or just to show code}
\usage{
run_cli()
}
\value{
logical
}
\description{
whether or not to run cli or just to show code
}
\keyword{internal}
