% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{run_step}
\alias{run_step}
\title{run_step}
\usage{
run_step(transform, type, inpaths = "", outpaths = "", ...)
}
\arguments{
\item{transform}{list. A named list of characters. The first element should
be the function name. All other elements optional parameters into the function.}

\item{type}{character. The type of transformation done for this step. Can
be one of five values: 
'0f1' - For functions with no input sets, creating one output set. 
'0fM' - For functions with no input sets, creating multiple output sets. Sets
must be returned in the same order that they're named in the pipeline.
'1f1' - For functions with one input set and one output set. 
'1fM' - For functions with one input set and multiple output sets. Sets must
be returned in the same order that they're named in the pipeline.
'Mf1' - For functions requiring multiple input sets, but producing only one 
output set. Sets must be input in the order that they will be passed into the
transformation function.
'MfM' - For functions with multiple inputs and multiple outputs. Input and 
output sets must be the same length. Element one of the input set produces
element one of the output set.}

\item{inpaths}{character. A vector of one or more of the set names used as
inputs to this step.}

\item{outpaths}{character. A vector of one or more of the set names created
by this step.}

\item{...}{list. A list of optional parameters to be set as local configs for
that particular step. Any config not specified in the step gets set to the
global values for that config.}
}
\value{
If \code{run_checks == FALSE}, the output is a list of the storage
paths created by the pipeline. If \code{run_checks == TRUE}, the output is
a dataframe listing each step, set created, and value for all check functions.
Only run if run_pipe = TRUE
Only run if run_checks = TRUE
Function run_all_checks will take in list of outpaths and list of checks
}
\description{
Processes a single step in the pipeline. Runs the step, if applicable. Runs
checks, if applicable.
}
