% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_base_url}
\alias{select_base_url}
\title{Select base URL
Helper function to switch base URLs depending on PIP server being used}
\usage{
select_base_url(server)
}
\arguments{
\item{server}{character: c("prod", "qa", "dev")}
}
\value{
character
}
\description{
Select base URL
Helper function to switch base URLs depending on PIP server being used
}
