% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cp_ki.R
\name{unnest_ki}
\alias{unnest_ki}
\title{Unnest the key indicators}
\usage{
unnest_ki(out)
}
\arguments{
\item{out}{parsed and simplified output from cp-key-indicators endpoint}
}
\value{
data frame, unnested.
}
\description{
Unnest the key indicators
}
\section{Functions}{
\itemize{
\item \code{unnest_ki()}: takes the simplified output from cp-key-indicators endpoint and unnests it.

}}
