% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{build_request}
\alias{build_request}
\title{build_request}
\usage{
build_request(
  country,
  year,
  povline,
  popshare,
  fill_gaps,
  group_by,
  welfare_type,
  reporting_level,
  version,
  ppp_version,
  release_version,
  format,
  server
)
}
\arguments{
\item{server}{}
}
\value{
httr2 request
}
\description{
build_request
}
