% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{call_aux}
\alias{call_aux}
\title{call a table from .pip env}
\usage{
call_aux(table = NULL)
}
\arguments{
\item{table}{character: name of table in .pip env. If NULL, it displays the
names of tables available in .pip env}
}
\value{
data frame of auxiliary table
}
\description{
call a table from .pip env
}
\examples{
# call one table

get_aux("gdp", assign_tb = TRUE, replace = TRUE)
call_aux("gdp")

# see the name of several tables in memory
tb <- c("cpi", "ppp", "pop")
lapply(tb, get_aux, assign_tb = TRUE, replace = TRUE)
call_aux()
}
