% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_grouped_stats_to_csv}
\alias{change_grouped_stats_to_csv}
\title{Change the list-output to dataframe (Function from pipapi)}
\usage{
change_grouped_stats_to_csv(out)
}
\arguments{
\item{out}{output from wbpip::gd_compute_pip_stats}
}
\value{
dataframe
}
\description{
Change the list-output to dataframe (Function from pipapi)
}
