% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{check_api}
\alias{check_api}
\title{Check internet connection and API status}
\usage{
check_api(api_version = "v1", server = NULL)
}
\arguments{
\item{api_version}{character: API version}

\item{server}{character: Server. For WB internal use only}
}
\value{
character
}
\description{
Check internet connection and API status
}
\examples{
\dontrun{
check_api()
}
}
