% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datt_rural}
\alias{datt_rural}
\title{Datt (1998) grouped data for rural india, 1983}
\format{
A data frame with 13 observations on the following 6 variables:
\describe{
\item{monthly_pc_exp}{Welfare range class}
\item{mean_monthly_pc_exp}{Mean welfare for given welfare range class}
\item{percentage_of_persons}{Percentage of individuals in given welfare class}
\item{L}{Cumulative welfare}
\item{p}{Cumulative population}
\item{area}{rural}
}

@source Datt, G. (1998). See get_cp vignette.
}
\usage{
datt_rural
}
\description{
Dataset from Datt (1998) with grouped data for rural India in 1983.
}
\keyword{datasets}
