% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datt_urban}
\alias{datt_urban}
\title{Grouped data for urban india, 1983}
\format{
A data frame with 13 observations on the following 6 variables:
\describe{
\item{monthly_pc_exp}{Welfare range class}
\item{mean_monthly_pc_exp}{Mean welfare for given welfare range class}
\item{percentage_of_persons}{Percentage of individuals in given welfare class}
\item{L}{Cumulative welfare}
\item{p}{Cumulative population}
\item{area}{urban}
}

@source Sarvekshana N26 Vol 9 N 4, and Datt, G. (1998) for methodology. See get_cp vignette.
}
\usage{
datt_urban
}
\description{
Dataset from Sarvekshana N26 Vol 9 N 4, created by the author following
Datt(1998) methodology with grouped data for urban India in 1983.
}
\keyword{datasets}
