% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_aux}
\alias{get_aux}
\title{Get AUX table}
\usage{
get_aux(table = NULL, api_version = "v1")
}
\arguments{
\item{table}{Aux table}
}
\value{
tibble
}
\description{
Get an auxiliary table. If no table is specified it a vector with possible
inputs will be returned.
}
\examples{
# Get list of tables
x <- get_aux()

# Get GDP data
df <- get_aux("gdp")

}
