% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_aux}
\alias{get_aux}
\title{Get AUX table}
\usage{
get_aux(
  table = NULL,
  version = NULL,
  api_version = "v1",
  format = c("json", "csv", "rds"),
  simplify = TRUE,
  server = NULL
)
}
\arguments{
\item{table}{Aux table}

\item{version}{character: Data version. See \code{get_versions()}}

\item{api_version}{character: API version}

\item{format}{character: Response format}

\item{simplify}{logical: If TRUE (default) the response is converted to a
data frame}

\item{server}{character: Server. For WB internal use only}
}
\value{
tibble or character vector
}
\description{
Get an auxiliary table. If no table is specified a vector with possible
inputs will be returned.
}
\examples{
# Get list of tables
x <- get_aux()

# Get GDP data
df <- get_aux("gdp")
}
