% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cache_info}
\alias{get_cache_info}
\title{Provides some information about cached items}
\usage{
get_cache_info()
}
\value{
character.
}
\description{
Provides some information about cached items
}
\examples{
\dontrun{get_cache_info()}
}
