% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_countries}
\alias{get_countries}
\title{Get countries}
\usage{
get_countries(
  version = NULL,
  api_version = "v1",
  format = c("json", "csv", "rds"),
  server = NULL
)
}
\arguments{
\item{version}{character: Data version. See \code{get_versions()}}

\item{api_version}{character: API version}

\item{format}{character: Response format}

\item{server}{character: Server. For WB internal use only}
}
\value{
tibble
}
\description{
Get a dataset with all countries used in PIP.
}
\examples{
get_countries()
}
