% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_countries}
\alias{get_countries}
\title{get_countries}
\usage{
get_countries(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\description{
Returns a table countries with their full names, ISO codes, and
associated region code
}
\examples{
\dontrun{
# Short hand to get countries
get_countries()
}
}
