% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_cpi}
\alias{get_cpi}
\title{get_cpi()}
\usage{
get_cpi(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\arguments{
\item{version}{character: Data version. See \code{get_versions()}}

\item{api_version}{character: API version}

\item{format}{character: Response format}

\item{server}{character: Server. For WB internal use only}
}
\description{
Returns a table of Consumer Price Index (CPI) values used for
poverty and inequality computations.
statistics
}
\examples{
\dontrun{
# Short hand to get cpi
get_cpi()
}
}
