% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_dictionary}
\alias{get_dictionary}
\title{get_dictionary}
\usage{
get_dictionary(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\description{
Returns a data dictionary with a description of all variables
available through the PIP API.
}
\examples{
\dontrun{
# Short hand to get dictionary
get_dictionary()
}
}
