% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd}
\alias{get_gd}
\title{Get grouped stats}
\usage{
get_gd(
  cum_welfare = NULL,
  cum_population = NULL,
  estimate = c("stats", "lorenz", "params"),
  requested_mean = 1,
  povline = 1,
  popshare = NULL,
  n_bins = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  simplify = TRUE,
  server = NULL
)
}
\arguments{
\item{cum_welfare}{numeric: Cumulative welfare values.}

\item{cum_population}{numeric: Cumulative population values.}

\item{estimate}{character: One of "stats", "lorenz", "params".}

\item{requested_mean}{numeric: Requested mean.}

\item{povline}{numeric: Poverty line. Required for endpoint = "grouped-stats".}

\item{popshare}{numeric: Proportion of the population living below the
poverty line}

\item{n_bins}{numeric: Number of bins. Required for endpoint = "lorenz-curve".}

\item{api_version}{character: API version}

\item{format}{character: Response format either of c("rds", "json", "csv")}

\item{simplify}{logical: If TRUE (the default) the response is returned as a
\code{tibble}}

\item{server}{character: Server. For WB internal use only}
}
\description{
Get grouped stats from the PIP API.
}
