% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_gdp}
\alias{get_gdp}
\title{get_gdp()}
\usage{
get_gdp(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\arguments{
\item{version}{character: Data version. See \code{get_versions()}}

\item{api_version}{character: API version}

\item{format}{character: Response format either of c("rds", "json", "csv")}

\item{server}{character: Server. For WB internal use only}
}
\description{
Returns a table of Growth Domestic Product (GDP) values used for
poverty and inequality statistics.
}
\examples{
\dontrun{
# Short hand to get gdp
get_gdp()
}
}
