% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_hfce}
\alias{get_hfce}
\title{get_hfce}
\usage{
get_hfce(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\arguments{
\item{version}{character: Data version. See \code{get_versions()}}

\item{api_version}{character: API version}

\item{format}{character: Response format}

\item{server}{character: Server. For WB internal use only}
}
\description{
Returns a table of Household Final Consumption Expenditure (HFCE) values
used for poverty and inequality computations.
}
\examples{
\dontrun{
# Short hand to get hfce
get_hfce()
}
}
