% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_interpolated_means}
\alias{get_interpolated_means}
\title{get_interpolated_means}
\usage{
get_interpolated_means(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\arguments{
\item{version}{character: Data version. See \code{get_versions()}}

\item{api_version}{character: API version}

\item{format}{character: Response format either of c("rds", "json", "csv")}

\item{server}{character: Server. For WB internal use only}
}
\description{
Returns a table of key information and statistics for all years
for which poverty and inequality statistics are either available (household
survey exists) or extra- / interpolated.
Please see \code{\link{get_dictionary}} for more information about
each variable available in this table.
}
\examples{
\dontrun{
# Short hand to get interpolated_means
get_interpolated_means()
}
}
