% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_pip_info}
\alias{get_pip_info}
\title{Get PIP info}
\usage{
get_pip_info(api_version = "v1", server = NULL)
}
\arguments{
\item{api_version}{character: API version}

\item{server}{character: Server. For WB internal use only}
}
\value{
list
}
\description{
Get information about the API.
}
\examples{
\dontrun{
get_pip_info()
}
}
