% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_pop_region}
\alias{get_pop_region}
\title{get_pop_region}
\usage{
get_pop_region(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\description{
Returns a table of total population by region-year. These values
are used for the computation of regional aggregate poverty statistics.
}
\examples{
\dontrun{
# Short hand to get pop_region
get_pop_region()
}
}
