% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_region_coverage}
\alias{get_region_coverage}
\title{get_region_coverage}
\usage{
get_region_coverage(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\description{
Return a table of regional survey coverage: Percentage of
available surveys for a specific region-year.
}
\examples{
\dontrun{
# Short hand to get region_coverage
get_region_coverage()
}
}
