% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux.R
\name{get_survey_means}
\alias{get_survey_means}
\title{get_survey_means}
\usage{
get_survey_means(
  version = NULL,
  api_version = "v1",
  format = c("rds", "json", "csv"),
  server = NULL
)
}
\description{
Returns a table of all available surveys and associated key
statistics. Please see \code{\link{get_dictionary}} for more information about
each variable available in this table.
}
\examples{
\dontrun{
# Short hand to get survey_means
get_survey_means()
}
}
