% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_versions}
\alias{get_versions}
\title{Get versions}
\usage{
get_versions(api_version = "v1", server = NULL)
}
\arguments{
\item{api_version}{character: API version}

\item{server}{character: Server. For WB internal use only}
}
\value{
character
}
\description{
Get available data versions.
}
\examples{
get_versions()
}
