% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_versions}
\alias{get_versions}
\title{Get versions}
\usage{
get_versions(api_version = "v1", server = NULL, simplify = TRUE)
}
\arguments{
\item{api_version}{character: API version}

\item{server}{character: Server. For WB internal use only}

\item{simplify}{logical: If TRUE (the default) the response is returned as a
\code{tibble}}
}
\value{
tibble or list
}
\description{
Get available data versions.
}
\examples{
\dontrun{
get_versions()
}
}
