% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_error_body}
\alias{parse_error_body}
\title{parse_error_body}
\usage{
parse_error_body(resp)
}
\arguments{
\item{resp}{}
}
\value{
character
}
\description{
Helper function to parse error messages generated by the PIP API
}
