% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_response}
\alias{parse_response}
\title{parse_response}
\usage{
parse_response(res, simplify)
}
\arguments{
\item{res}{A httr response}

\item{simplify}{logical: If TRUE (the default) the response is returned as a
\code{tibble}}
}
\value{
If \code{simplify = TRUE}, it returns a tibble with the requested content.
If \code{simplify = FALSE}, it returns a list of class "pip_api" with
information about the PIP API query
}
\description{
parse_response
}
\keyword{internal}
