% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pip_is_transient}
\alias{pip_is_transient}
\title{pip_is_transient}
\usage{
pip_is_transient(resp)
}
\arguments{
\item{resp}{A httr response}
}
\value{
logical
}
\description{
Helper function to determine if an error is due to the number of requests
going over the rate limit
}
