% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_aux}
\alias{set_aux}
\title{Set auxiliary table in .pip environment for later call}
\usage{
set_aux(table, value, replace = FALSE)
}
\arguments{
\item{table}{character: name of the table in .pip env}

\item{value}{data to be saved}

\item{replace}{logical.}
}
\value{
Invisible TRUE if set correctly. FALSE otherwise
}
\description{
Set auxiliary table in .pip environment for later call
}
\keyword{internal}
